Admin addressing changes applied/reflected at runtime.
The admin network is intended to be used between a subcloud
and system controller. Because the (existing) management network
is so embedded in other parts of the StarlingX system, it makes
it prohibitively hard to re-configure this network after initial
installation.  The admin network is intended to be isolated from
the management network, allowing re-configuration of the network
parameters in the case that the physical network between subcloud
and system controller has been changed.

For this feature, there is a requirement for admin subnet information
to be applied on the system without host lock/unlock.  This commit
provides this functionality by calling into a runtime puppet
manifest to apply the networking configuration, as well as update
SM, haproxy, and keystone endpoints.

Test Plan:
  PASS: Delete current admin addrpool, add new addrpool, assign to
    a network, and assign interface to network.  eg:
        system addrpool-delete $(system addrpool-list | grep admin \
            | awk '{print $2}')
        system addrpool-add --ranges 192.168.101.2-192.168.101.10 \
            --gateway-address 192.168.101.1 admin 192.168.101.0 24
        system network-add admin admin false \
            $(system addrpool-list | grep admin | awk '{print $2}')
        system interface-network-assign controller-0 enp0s3 admin
  PASS: Ensure after the runtime manifest runs (see associated review)
    that the IP addressing, haproxy, and endpoints are updated
    appropriately.

Story: 2010319
Task: 46926

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/868362

Signed-off-by: Mayank Patel <mayank.patel@windriver.com>
Co-Authored-By: Steven Webster <steven.webster@windrvier.com>
Change-Id: I121add8df3e77f02b29d9131645f880f78c0a4c5