Remove listener to monitor dcmanager keystone password
Because the dcmanager password is not referenced by the
cert-mon.conf in the subcloud, there's no need to update
anything during subcloud re-homing in the subcloud. This
commits removes keystone listener support to apply the
puppet manifest for dcmanager keystone password change.

Test passed:
Update the dcmanager's password, verify that the runtime
class for dcmanager did not get applied.

Story: 2010230
Task: 46943

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I11b1e4030f16b49a331f06f738d359d6a617042f