Optimize keystone_listener
In the old code, each user corresponded to a notification endpoint.
When the keystone password was changed, each user would get the same
notification. Therefore, the "info" method in the NotificationEndpoint
class and the "get_keystone_users" method in the OpenstackOperator class
would be called multiple times. This implementation may have race
condition risk and is hard to be debugged.

In the new code, there's only one NotificationEndpoint instance for
all of the users so that it can make sure no duplicated calls for
a single notification.

Test plan:
Verify - rehoming test for DX subcloud without host swact

Story: 2010230
Task: 46950

Signed-off-by: Li Zhu <li.zhu@windriver.com>
Change-Id: I77dcad070c0a825514ba0c971696bfaff8031177