Check against device_path and device_node on precheck upgrade.
During the precheck upgrade the function only check against the
device_path however there cases that boot_device or rootfs_device
on the i_host could be either /dev/sda or a by_path name.

This commit check it and prevents a not found issue also changes
the rootfs_device to boot_device to be consistent with migration
scripts.

Closes-bug: 1997889

Test Plan:
PASS: Precheck works in worker node with boot_device as /dev/sda
PASS: Precheck works in other hosts with boot_device as by_path

Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>
Change-Id: Ic0a94039fa567904e104a6dba6195f0f80ce3b38