Fix hung up upgrade on activating-hosts state
During upgrade-activate, if hosts get remain in config out-of-date
status for some reason, the upgrade gets stuck on the status
activating-hosts, not allowing the user to solve the activation
issues and re-execute upgrade-activate.

This commit adds a timeout logic within the upgrade status audit,
such that when it runs, if there are hosts config out-of-date,
it calculates how much time has passed since the upgrade manifests
were triggered and fails the upgrade activate if this value exceeds
a determined timeout value.

Test Plan
PASS: successfully upgrade AIO-DX
PASS: force config out-of-date stuck state during upgrade-activate
      and verify that upgrade is set to activation-failed after the
      timeout period has passed

Closes-bug: 1998861

Change-Id: I40eb1e8be628ffacf034f92fb907fef2b48f2505
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>