Allow FluxCD to be upgraded from N to N+1
With the introduction of [1] FluxCD needs to be updated.
Enhancing upgrade-fluxcd-controllers.yml to allow FluxCD
upgrades is done part of [2].

When upgrading from 22.06 we need the flux manifest to be reapplied
(and pods will be restarted part of the new deployment).
This is designed future proof, upgrading from generic N to N+1 would
work.

Tests on AIO-SX:
Using the load for [2], emulated upgrades call.
PASS: Observed no pod restart when emulating upgrade from 21.12 as there
should be no pod in the first place when coming from 21.12
/etc/upgrade.d/62-enable-fluxcd-controllers.py 21.12 22.12 activate
PASS: Observed pods restart picking the new manifests whe emulating
upgrade from 22.06
/etc/upgrade.d/62-enable-fluxcd-controllers.py 22.06 22.12 activate

[1]: https://review.opendev.org/c/starlingx/ansible-playbooks/+/866820
[2]: https://review.opendev.org/c/starlingx/ansible-playbooks/+/867032
Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/867032
Partial-Bug: 1999032
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I14c7802ef405a2ea4665b2a2b64e5daaf0eef009