Only use required fault management endpoint
When managing a large number of subclouds on a DC System Controller,
sysinv-api freezes and the controller swacts.
This happens due to sysinv-api RequestContext service catalog having
the fm endpoints of all subclouds, which is too large for sysinv to
handle when using the ZeroMQ RPC backend.

This change filter the RequestContext service catalog to only store
the required endpoint, instead of the endpoints of all subclouds.

TEST PLAN:
PASS: AIO-SX: Install, bootstrap and host-unlock
PASS: DC: Edit the source files with this change, restart sysinv-api
      service, then install and manage a large number of subclouds
PASS: AIO-SX/DC: Perform system commands that interacts with fm
      (system health-query, for instance)

Story: 2010087
Task: 46444

Signed-off-by: Alyson Deives Pereira <alyson.deivespereira@windriver.com>
Change-Id: I0a8baad3022a64c25e188ac6f6da24548605785a