Enable add ceph-mon on any compute without storage-0 host
When deploying a Standard setup using a mgmt network with range to
match exacly the number of required nodes, it does not accept to
add ceph-mon on compute-0 node because it tries to reserve an IP
address to storage-0 which will never be configured. Nonetheless,
there is no available ip address to reserve for storage-0 and
system ceph-mon-add command fails.

To avoid reserving ip address to a host that will never be installed
now it verifies if it is trying to add ceph-mon to storage or worker
and call the right function to reserve the IP address.

Test Plan:
  PASS: Deploy Standard setup (2+3) with mgmt network netmask 125 on
        IPv6 network and after adding all 5 hosts, add ceph-mon on
        compute-0 and unlock all workers nodes.
  PASS: Deploy Dedicated Storage setup (2+1+2) with mgmt network
        netmask 29 on IPv4 network.

Closes-bug: 1999286

Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>
Change-Id: I1def05f4580131ea6b5a6202461a2fa921d8e5f2