Reduce the number of partition DB updates
When the partition code receives a partition structure
from sysinv-agent, the fields are strings.
The database fields are integers, so the comparison
always considers the values to be mis-matched and attempts
to update the database.

This can cause stale 'size' values from the agent query
to overwrite the DB during a partition resize operation.

Test Plan:
  PASS: create and update several partitions
  PASS: verify that the 'PART update' log is not created
once for every partition when doing a partition update.

Closes-Bug: #1999586
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I3113e60a09951aa0efc5a6e549b1614976f75f35