Cert Manager not upgraded in certain upgrade paths
Cert Manager is excluded in 65-k8s-app-upgrade.sh for
upgrade from 21.12 to 22.12, but should be included for 22.06
to 22.12 upgrade. This is because the dedicated cert manager upgrade
script 64-upgrade-cert-manager.sh handles the 21.12 to 22.12
upgrade, and does not run for the 22.06 to 22.12 upgrade.

Test Cases:

PASS: Perform platform upgrade on a single node system. Ensure
      cert-manager is running the new version after upgrade-activate

Closes-Bug: 1999801
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>
Change-Id: Ib77c6f84900c00697654035585f4888e2525a40c