Update database if kubelet upgrade already done
If something goes wrong during the kubelet upgrade, it's possible
for the kubelet upgrade to complete but for the sysinv database
to be set to the "upgrading-kubelet-failed" state. If we get
into this state and recover kubelet such that sysinv can determine
that it has been upgraded, then on a retry we should just update
the sysinv database rather than reject the kubelet upgrade.

Test Plan:
PASS: verify kubelet upgrade status transitions from
upgrading-kubelet-failed to upgraded-kubelet. Manually edit sysinv db
kubelet-upgrade-status field and re-issue "system kube-host-upgrade
controller-0 kubelet".

Closes-Bug: 1999877

Signed-off-by: Sachin Gopala Krishna <saching.krishna@windriver.com>
Change-Id: I4cba7d138b353b6182706ef0288a262a48fdb726