Fix validation for latitude and longitude params.
System modify command accepts garbage input for latitude and longitude.
The old validation was to check if either latitude or longitude had
more than 30 characters. It was added a regex for latitude and
longitude, allowing only numbers between -90 and +90 for
latitude and numbers between -180 and +180 for longitude.

Test Plan:
Pass: AIO-SX full deployment.
Pass: run system modify with a latitude greater than +90.
Expected a message for invalid input.
Pass: run system modify with a latitude between -90 and +90.
Expected the command to run succesfully.
Pass: run system modify with a non-numeric string as latitude.
Expected a message for invalid input.
Pass: run system modify with a longitude greater than +180.
Expected a message for invalid input.
Pass: run system modify with a longitude between -180 and +180.
Expected the command to run succesfully.
Pass: run system modify with a non-numeric string as longitude.
Expected a message for invalid input.

Closes-Bug: 2001677

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I668f158086397a1d155ad5ded2e7fa75d1e9ffde