System mode modify fails for duplex systems
The system modify command was not working properly to update
the system_mode on duplex systems, so updating from duplex to
duplex-direct and vice versa was not possible. Changes were
made to ensure that this modification could be made
while keeping the SX to DX update rules implemented.
Additionally, the same rules applied during simplex to duplex
migration were also applied during simplex to duplex-direct
migration.

Automated tests were also added to make sure future changes
don't change the behaviour of the system without warnings.

Test plan:
PASS - Install and bootstrap SX, migrate it to DX and
duplex-direct and check if the system modify command is
working properlly on CentOS.
PASS - Install and bootstrap SX, migrate it to DX and
duplex-direct and check if the system modify command is
working properlly on Debian.

Closes-bug: 2002561

Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>
Change-Id: I42cff0291945a42b10fa9a4abecaf3c936ffc7e6