Update severity check and cert-alarm override
Updating check for user input in severity value, and alarm
enabled/disabled. For the severity value updating the check
and Logging when an user override exist. In case of alarm
enabled/disabled changing the conditions to clear the existing
alarms for the certificate when the user disable the alarm.

Test Plan:

PASS: Add user override in the certificate and check if the severity
      changes in the cert-alarm alarm.
PASS: Add user override and check for the custom log warning that the
      cert-alarm will use now a user custom value for the severity.
PASS: Disable the alarm for the certificate and check if all related
      alarms have been cleared.
PASS: Change the field back to Enabled and check if all alarms is
      raised.

Closes-Bug: 2002823
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>
Change-Id: I032ef6e880fbede4422df360fdf560602fec95f8