Update network interface name length limit, change VLAN OS interface naming rule, refactor tests
This commit changes the interface name length limit to a global value
of 15, which is the maximum value allowed by the kernel. Previously
it was 10 to allow for the 5 remaining characters to be used as VLAN
ID, but this restriction does not work well, since there are systems
where Linux generate names with more than 10 characters for some
underlying ethernet interfaces, causing VLAN interface creation to
fail anyway. Instead, the OS interface naming convention for VLANs
was changed to be the following:

- For platform interfaces, the interface name for the Linux OS
  continues to be vlanNNNN, where NNNN is the VLAN ID number.

- For non-platform interfaces, the interface name for the Linux OS
  will be the same as the sysinv interface name (ifname)

- For non-platform interfaces, if the sysinv interface name follows
  the two patterns that are recognized as VLAN definitions when put
  in /etc/network/interfaces (vlanNNNN and *.NNNN, like vlan6453 of
  myinterface.1234), the OS name will be:
  > For the vlanNNNN pattern, vlan#NNNN
  > For the interface.NNNN pattern, interface#NNNN
  This is necessary because such patterns cause ifup to deduce the
  underlying interface and VLAN ID from the name, and they may differ
  from the ones specified in sysinv.

This is achieved by the following changes:

- Add code to handle non-platform interfaces whose name follow the
  patterns which are recognizable by ifup. Patterns like vlanNNN are
  replaced by vlan#NNN, and the ones like xxx.NNN by xxx#NNN. Hashtag
  (#) character was chosen because it's accepted by Linux, but not by
  sysinv, so it's impossible to generate name collisions. If
  underscore was chosen, duplicate names might occur (for example,
  in a host with two distinct interfaces named myif_555 and myif.555,
  for both of them the OS name would be myif_555, causing a name
  collision).

- Add the command 'ip link add link <eth_port> name <ifname> type
  vlan id <vlan_id>' to the pre-up operation for non-platform
  interfaces. Since the name does not follow a recognized VLAN
  pattern, there has to be an explicit instruction to create the
  interface.

Also, the unit tests for api/interfaces in which fails were expected
did not have validations for the error messages, causing 7 of the
tests to be wrongly evaluated. This commit adds the proper checks for
error messages in the test functions, and enforcements in the helper
methods. This way, all existing tests are evaluated correctly,
and any new written tests will fail if the expected error messages
are not specified.

Test plan

PASS Create a VLAN interface using a name with 15 characters (ok)
PASS Create a VLAN interface using a name with 16 characters (fail)
PASS Check OS interface names for non-platform interfaces whose names
     follow ifup VLAN patterns
PASS Boot host and check for proper VLAN interfaces creation
PASS Install duplex-direct system

Closes-Bug: #2003589

Signed-off-by: Lucas Ratusznei Fonseca
               <lucas.ratuszneifonseca@windriver.com>

Change-Id: I73594f21fc8e0e0ad27a40e21fbb6b7a4c3347d6