DC: Set system controller gateway for admin network
When a subcloud is provisioned with the admin network,
it is required to set the system controller gateway
address so that a route to the system controller is
created when the distributed cloud config is run.

This commit also fixes a bug in the handling of
'address already exists' exception messaging during
route creation.

Test Plan:

Provision a subcloud to use the admin network
  - Ensure a route to the system controller is created
  - Ensure the subcloud can become managed, in-sync

Provision a subcloud to use the mgmt network (regression)
  - Ensure a route to the system controller is created
  - Ensure the subcloud can become managed, in-sync

Story: 2010319
Task: 46910

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I0c62fbecd90258213075f2c22af60529d9ed2b88