Update state available after size is finalized
Currently, controller fs state is updated for
only resize commands are executed.
But the initial configuration does not
match this case, and the state stays "None".
(The only problem is state in db for initial
configuration. Controller fs itself is
usable and properly configured.)

All fs states should be "available" including
the initial configuration if the fs size is
finalized.
This fix is to add fs to resized list if the
state is None because the initial configuration
is done, so that the state is set "available"
properly later.

Closes-bug: 2004564

Test Plan:
PASS: Fresh install (SX) and confirm all
      fs states are "available"
      (system controllerfs-list)
PASS: Fresh install (DX) and confirm
      - All fs states are "available"
        (system controllerfs-list)
      - drbd-oveview shows proper role,
        connected and synced
      - No drbd fs related alarm
      - No error in system host-swact
PASS: After fresh install (SX), modifies
      fs size (system controllerfs-modify)
      and confirm the state;
      - "drbd_fs_resizing_in_progress"
        during processing
      - "available" after resized
      - "drbd-overview" shows the
        resized size
PASS: After fresh install (DX), modifies
      fs size (system controllerfs-modify)
      and confirm the state;
      - "drbd_fs_resizing_in_progress"
        during processing
      - "available" after resized
      - "drbd-overview" shows the
        resized size, proper role,
        connected and synced

Signed-off-by: Takamasa Takenaka <takamasa.takenaka@windriver.com>
Change-Id: I88921041bc850634967587d3f811dfc1e22fd8bd