Update not ssl_ca certificate removal message
The uninstall of certificates that are not of the type ssl_ca is not
supported by design, instead installing a new certificate should be
used to update the certificate. This fix provides an error message
that suggests the correct action for performing the certificate
replacement.

Test Plan:
PASS: In a simplex, install a certificate that is not of type ssl_ca
      (e.g. openstack_ca) and perform the command "system certificate-
      uninstall <cert uuid>" and the error message should present the
      information about the correct update method for the type of
      certificate that are not supported for certificate-uninstall.
      The updated message comes from the client which handles the
      error preemptively in this scenario.

PASS: In order to test the API, using the same configuration as the
      previous scenario, perform a request to the sysinv REST API
      using the DELETE method and check the payload. The same error
      message shown by the client should be found.


Closes-Bug: 2004662
Signed-off-by: Alan Bandeira <Alan.PortelaBandeira@windriver.com>
Change-Id: Idf3d8da2c43e23d821925675ecc6af6eb8126eaa