adjust oidc-auth-apps upgrade logic
oidc-auth-apps in 21.12 is from a different provider and not compatible
with the application in 22.06 and 22.12.

The upgrade from 22.06 to 22.12 is handled by 65-k8s-app-upgrade.sh.
Upgrade from 21.12 is handled by another script.

Test Plan:
PASS - Build master branch
PASS - Unit test logic of conditions with variables for cert-manager and
       oidc-auth-apps, 21.12, 22.06, 22.12
PASS - manually apply the change for 21.12 to 22.12 upgrade
PASS - manually apply the change for 22.06 to 22.12 upgrade

Story: 2009303
Task: 47367

Change-Id: I189e1e42f55013ab357f55a17459c6c02becb1f7
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>