Add mpath support to validate_platform_backup
The command 'system health-query-upgrade' returns the value [Fail]
for the field 'Valid platform-backup partition',
that happens because the lsblk command is not able to return
a device path when multipath disks are used or when the root
partition is under cgts--vg.

To fix this, backup commands were added in case lsblk fails to
return a valid path.

Test Plan:
PASS: Upgrade AIO-SX (with and without multipath):
      Valid platform-backup partition: [Ok]
      on "system health-query-update"
PASS: Upgrade AIO-DX (with and without multipath):
      Valid platform-backup partition: [Ok]
      on "system health-query-update"
PASS: Upgrade Standard (with and without multipath):
      Valid platform-backup partition: [Ok]
      on "system health-query-update"

Story: 2010608
Task: 47648

Signed-off-by: Matheus Guilhermino <matheus.machadoguilhermino@windriver.com>
Change-Id: I2655dd4cfe1c91a8bcd2d7d8e10ddb662a58d6ce