Remove metadata from apps_metadata dict when deleting app
When some application is deleted, it's information remains
cached in all dicts and lists from apps_metadata, with this
structure:

    apps_metadata =
        {apps: {},
        platform_managed_apps_list: {},
        desired_states: {},
        ordered_apps: []}

The mentioned behavior was causing a bug after an upgrade from
stx 6 to stx 8, because an app that has no support in stx 8 is
deleted but continue cached in apps_metadata, and when
k8s_application_audit() is run after the upgrade is complete,
the system uses the cached list to try to upload this app until
the upload-failed state appears in the alarms list.

Now the created method is called right after deleting an
application, this way the app will be completely out of the
system.

Test-Plan:
  PASS: Ugrade from stx 6.0 to 8.0 in a standard config lab.
  PASS: During the upgrade, an app from the previous version
  which has no support in the next version is automatically
deleted.
  PASS: Validate that the deleted app is not in any
  collection from apps_metadata.
  PASS: Validate that the upload try of the deleted app doesn't
happen anymore.

Closes-Bug: 2009025

Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>
Change-Id: I6a54218b398493acc931c5eca34b800383b16cc0