AppFwk: Enable auto update for non platform managed apps
We have a gap in design where we want apps to be auto updated if
specific metadata is present in the app on the N+1 side. Metadata
represented by [1].

The gap in design is that the N version must also contain specific
metadata to register the app as 'platform managed app'. Metadata
represented by [2].

Tests on AIO-SX:
PASS: apply ptp-notification app in a patch, then restart sysinv
      observed the newer version of ptp-notification is auto-applied

[1]: https://opendev.org/starlingx/app-security-profiles-operator/src/commit/31b3f77eeefdc0be4c9c1f53c4e72888313a8d0e/stx-security-profiles-operator-helm/stx-security-profiles-operator-helm/files/metadata.yaml#L6-L7
[2]: https://opendev.org/starlingx/platform-armada-app/src/commit/4c765c35500da92d920d3fbf3c79b4cf02d4712b/stx-platform-helm/stx-platform-helm/files/metadata.yaml#L5-L6
Closes-Bug: 2009187
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Ia6f5646e78b8ed012e7dcb6ac7fa4c513f382bc2