Fix query_pci_id argument syntax
This change fixes the query_pci_id argument usage for custom ELF files.
The new argument was added [1] but tested only in isolation for the
query_pci_id.
Sysinv Agent is currently using the wrong syntax that needs to be "-e"
OR "--elfbinary=". Wrong argument usage on this script call will always
fail for sysinv and return code 2, instead of 0 for interfaces that
support DPDK or 1 for interfaces that do not support DPDK.

[1] https://review.opendev.org/c/starlingx/config/+/872979

TEST PLAN:
PASS - Build sysinv packages
PASS - Generate a new StarlingX ISO
PASS - Deploy an AIO-SX and customize the ELF target
      * For a valid ELF (e.g. /usr/bin/dpdk-testpmd), verify that the
        sysinv call to query_pci_id is no longer receiving return code 2

Closes-Bug: 2009487

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: Iec8c925d728be8d22f72bac3c8141904a051384b