sysinv: ignore bootstrap config file if missing
In commit cf94bebd9 we added a routine to edit the saved bootstrap
config file.  Originally this was fine since the file was guaranteed
to exist.

However, we now want to be able to do a "skip version" upgrade from
earlier releases where the bootstrap config file won't exist.  In this
case we should just ignore the missing file.

Accordingly, catch the specific error case and emit an informational
log rather than failing the K8s upgrade.

Closes-Bug: 2009619

Change-Id: I103f9b769a4803f85624a8fa4014ae11fa4cf227
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>