Support old Barbican registry secrets
Older versions of StarlingX created Barbican registry secrets as
"text/plain", that are retrieved as "string" in Python 3. This change
adds support to these secrets, that may be used in upgraded systems. In
non-upgraded newer systems, these secrets are created as
"application/octet-stream", being retrieved as "bytes" to be decoded
using UTF-8 in Python 3.

Test Plan:

PASS: Check that the system can retrive the registries usernames and
passwords when the secrets have the content type as "text/plain".

PASS: Check that the system can retrive the registries usernames and
passwords when the secrets have the content type as
"application/octet-stream".

Closes-Bug: 2009631
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I5a71239b09ef1124449dc66f86ef790e1f23222c