Removing reading file process in a subprocess
Details: This commit is to enhance the execution of lldpcli
in the subprocess.

The existing way is reading the OS release file during subprocess
execution. This file reading requires an extra file descriptor.

The commit is to only take "json0" value for Debian OS as
moving forward only Debian OS is supported.

Test Plan:

PASS: upgrade from 21.12 to 22.12 with this commit
PASS: fresh installation with this commit

Closes-Bug: 2009962

Signed-off-by: Junfeng (Shawn) Li <junfeng.li@windriver.com>
Change-Id: I1b35ab07775bbaee14a8c20513c398efcfa1d6e7