Admin network address pool robustness
Normally, a platform network's address pool is created at bootstrap
and is unable to be deleted after the initial configuration is
complete.

An exception to this is the admin network / address pool which is
able to be deleted and re-configured by a user at any time.

This facilitates the ability for a subcloud to be updated or
re-homed to adapt to network changes that may occur between
itself and the desired system controller.

Currently, most platform network address pools are created at
bootstrap by only specifying the range of addresses based on
the start-end addresses. When the address pool is associated
with a network, the appropriate floating, controller-0, and
controller-1 addresses are allocated.

In the case of the admin network, a user may decide to delete
the network/address-pool and re-create it with the (valid)
specification of a floating-address, controller-0-address,
and controller-1-address. ie:

system addrpool-add \
  --floating-address 192.168.103.2 \
  --controller0-address 192.168.103.3 \
  --controller1-address 192.168.103.4 \
  --gateway-address 192.168.103.1 \
  admin 192.168.103.0 24

Currently, this would cause the floating IP to be
allocated as 192.168.103.5, controller-0 IP to be allocated
as 192.168.103.6 and controller-1 IP to be allocated as
192.168.103.7 when the address pool is associated with
a network. This would not be what the user expects.

In addition, it was found that if a user does use the
options for floating-ip, controller0/1 ip, the current
code has some significant bugs:

1. Possible to have mixed IPv4/IPv6 addresses
2. Possible to specify floating/controller addresses
   as the network (subnet) address or the broadcast
   address
3. Possible to specify the floating/controller addresses
   completely outside the specified subnet.

This commit ensures that any user specification for the
floating or controller addresses are preferred over
auto-allocation.

Test Plan:

 The addition of the addrpool unit tests provide over 40
 new cases across IPv4/IPv6

 Pass: Bootstrap a subcloud using the admin network.
 Pass: On the subcloud, delete the admin addrpool and
       re-create it with various combinations of
       subnet/floating/controller-0/1 IP.  These
       combos are also covered by the unit tests.

Story: 2010319
Task: 46910

Change-Id: Ie59c26ae2a57b9cb570c7b20e3b40aa0f14fd95d